bin: change auth.sh to use template confs

+++ docker-run.sh
@@ -1,45 +1,43 @@
-#!/bin/bash
+#!/bin/bash -x
 
 ################################## ################################   ####  # ##
-# >> DOCKER-RUN-GLFTPD-WEBGUI
+# >> DOCKER-RUN-GLFTPD-V3 :: WEBUI
 ################################## ################################   ####  # ##
+# ENVIRONMENT VARIABLES:
+#
+# GLDIR="<path>"                   basedir for bind mounts (default=./glftpd)
+#                                  ( to reuse existing install set to /glftpd )
+# WEBUI_LOCAL=1                    run commands on same host, no docker [0|1]
+# WEBUI_AUTHMODE="basic"           auth mode [basic|glftpd|both|none]
+# NETWORK="host"                   docker network mode [host|bridge]
+# FORCE=1                          remove any existing container first [0|1]
+#
+# WEBUI_ARGS+= " --any-other-flags "      add any other docker run options
+#
+###################################################################   ####  # ##
 
-GLDIR=/glftpd
+#DEBUG=0
+
+GLDIR="./glftpd"
 GLFTPD=0
-LOCAL=1
 
-#NETWORK="$(docker network ls --format '{{.Name}}' --filter 'Name=shit')"
+WEBUI=1
+WEBUI_LOCAL=1
+WEBUI_AUTHMODE="basic"
+#NETWORK="host"
 
-# shellcheck disable=SC2288
-WEBUI_ARGS+= " --env WEBUI_AUTHMODE=none "
-
-#GLFTPD_ARGS="$*"
-WEBUI_ARGS=""
-NETWORK_ARGS="--network shit"
-GLFTPD_ARGS+=" $NETWORK_ARGS "
-#WEBUI_ARGS+=" $NETWORK_ARGS "
-
-#WEBUI_ARGS+=" --volume ${GLDIR:-/glftpd}:/glftpd "
-
-if [ "${LOCAL:-0}" -eq 1 ]; then
-  WEBUI_ARGS+=" $NETWORK_ARGS "
-  WEBUI_ARGS+=" --volume ${GLDIR:-/glftpd}:/glftpd "
-else
-  WEBUI_ARGS+="--network host"
-fi
-
-WEBUI_ARGS+=" --volume /var/run/docker.sock:/var/run/docker.sock "
-RM=1
-
+DOCKER_REGISTRY="ghcr.io/silv3rr"
 DOCKER_IMAGE_GLFTPD="docker-glftpd:latest"
 DOCKER_IMAGE_WEBUI="docker-glftpd-web:latest"
-DOCKER_REGISTRY="ghcr.io/silv3rr"
 
+GLFTPD_ARGS+="$*"
+WEBUI_ARGS+="$*"
+
+REMOVE_CT=1
 SCRIPTDIR="$(dirname "$0")"
 
 # get external/public ip
 if [ -z "$IP_ADDR" ]; then
-  # TODO: ip route show to exact 0/0 
   GET_IP="$( ip route get "$(ip route show 0.0.0.0/0 | grep -oP 'via \K\S+')" | grep -oP 'src \K\S+' )"
   IP_ADDR="${GET_IP:-127.0.0.1}"
 fi
@@ -53,6 +51,11 @@ BOT_STATUS="$(
   docker image inspect --format='{{ index .Config.Labels "gl.sitebot.setup" }}' "$DOCKER_IMAGE_GLFTPD" \
     2>/dev/null
 )"
+
+#WEBUI="$(
+#  docker image inspect --format='{{ index .Config.Labels "gl.web.setup" }}' "$DOCKER_IMAGE_GLFTPD" \
+#    2>/dev/null
+#)"
 
 if [ -s "$SCRIPTDIR/customizer.sh" ]; then
   IP_ADDR=$IP_ADDR ZS_STATUS=$ZS_STATUS BOT_STATUS=$BOT_STATUS \
@@ -68,40 +71,37 @@ echo "----------------------------------------------"
 echo "DOCKER-GLFTPD-RUN-V3"
 echo "----------------------------------------------"
 
-# select image
+# set runtime docker args
 
-LOCAL_IMAGE=$(
-  docker image ls --format='{{.Repository}}' --filter reference="$DOCKER_IMAGE_GLFTPD"
-)
-LOCAL_FULL_IMAGE=$(
-  docker image ls --format='{{.Repository}}' --filter reference="${DOCKER_IMAGE_GLFTPD/%:latest/:full}"
-)
-if [ -n "$LOCAL_IMAGE" ] && [ "${USE_FULL:-0}" -eq 0 ]; then
-  echo "* Found local image 'docker-glftpd'"
-elif [ -n "$LOCAL_FULL_IMAGE" ]; then
-  DOCKER_IMAGE_GLFTPD="${DOCKER_IMAGE_GLFTPD/%:latest/:full}"
-  echo "* Using full docker image ${LOCAL_FULL_IMAGE:-""})"
-else
-  # check if we already have 'full' tagged image, then keep using it
-  REGISTRY_FULL_IMAGE=$(
-    docker image ls --format='{{.Repository}}' --filter reference="${DOCKER_REGISTRY}/${DOCKER_IMAGE_GLFTPD/%:latest/:full}"
-  )
-  if [ -n "$REGISTRY_FULL_IMAGE" ] || [ "${USE_FULL:-0}" -eq 1 ]; then
-    DOCKER_IMAGE_GLFTPD="${DOCKER_IMAGE_GLFTPD/%:latest/:full}"
-  fi
-  echo "* Pulling image '${DOCKER_IMAGE_GLFTPD}' from registry '$DOCKER_REGISTRY'"
-  DOCKER_IMAGE_GLFTPD="${DOCKER_REGISTRY}/$DOCKER_IMAGE_GLFTPD"
-  docker pull "$DOCKER_IMAGE_GLFTPD"
+if [ "${DEBUG:-0}" -eq 0 ]; then
+  GLFTPD_ARGS+=" --detach "
+  WEBUI_ARGS+=" --detach "
 fi
 
-# set runtime docker args
+#WEBUI_ARGS+=" --add-host glftpd:127.0.0.1 "
+
+if [ -z "$NETWORK" ]; then
+  DOCKER_NETWORK="$(docker network ls --format '{{.Name}}' --filter 'Name=shit')"
+  if [ -n "$DOCKER_NETWORK" ] && [ "$DOCKER_NETWORK" = "shit" ];  then
+    NETWORK="shit"
+  fi
+fi
+
+if [ "${WEBUI_LOCAL:-0}" -eq 1 ]; then
+  WEBUI_ARGS+=" --mount type=bind,src=${GLDIR:-/glftpd},dst=/glftpd "
+  WEBUI_ARGS+=" --env WEBUI_PORT=4444 "
+  echo "* Running webui on host network: https://localhost:4444"
+else
+  WEBUI_ARGS+=" --publish "${IP_ADDR:-127.0.0.1}:4444:443" "
+fi
 
 # set max open files to prevent high cpu usage by some procs
 GLFTPD_ARGS+=" --ulimit nofile=1024:1024 "
 WEBUI_ARGS+=" --ulimit nofile=1024:1024 "
 
+# mount glftpd.conf
 if [ "${GLFTPD_CONF:-0}" -eq 1 ] || [ "${ZS_STATUS:-0}" -eq 1 ]; then
-  RM=0
+  REMOVE_CT=0
   if [ -d glftpd/glftpd.conf ]; then
     rmdir glftpd/glftpd.conf 2>/dev/null || { echo "ERROR: \"glftpd.conf\" is a directory, remove it manually"; }
   fi
@@ -112,7 +112,7 @@ if [ "${GLFTPD_CONF:-0}" -eq 1 ] || [ "${ZS_STATUS:-0}" -eq 1 ]; then
 fi
 
 if [ "${GLFTPD_CONF:-0}" -eq 1 ]; then
-  RM=0
+  REMOVE_CT=0
   echo "* Set docker ip:port"
   #GLFTPD_PASV_PORTS="$(sed -n -E 's/^pasv_addr (.*)/\1/p' glftpd/glftpd.conf)"
   if grep -Eq "^pasv_ports.*" glftpd/glftpd.conf; then
@@ -125,7 +125,7 @@ if [ -n "$GLFTPD_PASSWD" ]; then
 fi
 
 if [ "${GLFTPD_PERM_UDB:-0}" -eq 1 ]; then
-  RM=0
+  REMOVE_CT=0
   GLFTPD_ARGS+=" --mount type=bind,src=${GLDIR:-/glftpd}/ftp-data/users,dst=/glftpd/ftp-data/users "
   GLFTPD_ARGS+=" --mount type=bind,src=${GLDIR:-/glftpd}/ftp-data/groups,dst=/glftpd/ftp-data/groups"
   GLFTPD_ARGS+=" --mount type=bind,src=${GLDIR:-/glftpd}/etc,dst=/glftpd/etc "
@@ -133,14 +133,14 @@ fi
 
 # shellcheck disable=SC2174
 if [ "${GLFTPD_SITE:-0}" -eq 1 ]; then
-  GLFTPD_ARGS+=" --volume ${GLDIR:-/glftpd}/site:/glftpd/site:rw "
+  GLFTPD_ARGS+=" --mount type=bind,src=${GLDIR:-/glftpd}/site,dst=/glftpd/site:rw "
   WEBUI_ARGS+=" --mount type=bind,src=${GLDIR:-/glftpd}/site,dst=/app/glftpd/site "
 else
-  WEBUI_ARGS+=" --mount type=tmpfs,dst=/app/glftpd/site/NO_BIND_MOUNT " 
+  WEBUI_ARGS+=" --mount type=tmpfs,dst=/app/glftpd/site/NO_BIND_MOUNT "
 fi
 
 if [ "${BOT_STATUS:-0}" -eq 1 ]; then
-  RM=0
+  REMOVE_CT=0
   GLFTPD_ARGS+=" --mount type=bind,src=${GLDIR:-/glftpd}/sitebot,dst=/glftpd/sitebot "
   GLFTPD_ARGS+=" --publish ${IP_ADDR}:3333:3333 "
   for i in glftpd/sitebot/eggdrop.conf glftpd/sitebot/pzs-ng/ngBot.conf ; do
@@ -156,61 +156,103 @@ if [ "${BOT_STATUS:-0}" -eq 1 ]; then
   fi
 fi
 
+WEBUI_ARGS+=" --mount type=bind,src=/var/run/docker.sock,dst=/var/run/docker.sock "
+WEBUI_ARGS+=" --env WEBUI_AUTHMODE=${WEBUI_AUTHMODE:-basic} "
+
+# custom glftpd commands
+
 if [ -d entrypoint.d ]; then
-  RM=0
+  REMOVE_CT=0
   GLFTPD_ARGS+=" --mount type=bind,src=$(pwd)/entrypoint.d,dst=/entrypoint.d "
   echo "* Mount 'entrypoint.d' dir for custom commands"
 fi
 
 if [ -d custom ]; then
-  RM=0
+  REMOVE_CT=0
   if find custom/* >/dev/null 2>&1; then
     GLFTPD_ARGS+=" --mount type=bind,src=$(pwd)/custom,dst=/custom "
     echo "* Found files in 'custom', mounting dir"
   fi
 fi
 
-if [ "${RM:-1}" -eq 1 ]; then
+if [ "${REMOVE_CT:-1}" -eq 1 ]; then
   GLFTPD_ARGS+=" --rm  "
+  WEBUI_ARGS+=" --rm  "
 fi
 
-# remove existing containers which use local and/or registry images
+# remove existing container(s) which use local and/or registry images
 
 #set -x
-if [ "${WEBUI:-0}" -eq 1 ]; then
-  REGEX="( ${DOCKER_IMAGE_GLFTPD:-'docker-glftpd'}|${DOCKER_REGISTRY}/docker-glftpd)$"
-else
-  REGEX="(glftpd|glftpd-web|( ${DOCKER_IMAGE_GLFTPD:-'docker-glftpd'}|${DOCKER_IMAGE_WEBUI:-'docker-glftpd-web'}|${DOCKER_REGISTRY}/docker-glftpd.*))$"
+
+REGEX_GLFTPD="(glftpd| ${DOCKER_IMAGE_GLFTPD:-'docker-glftpd'}|${DOCKER_REGISTRY}/docker-glftpd)$"
+REGEX_WEBUI="(glftpd-web| ${DOCKER_IMAGE_WEBUI:-'docker-glftpd-web'})$"
+
+if [ "${GLFTPD:-0}" -eq 1 ]; then
+  REGEX="$REGEX_GLFTPD"
 fi
-docker ps -a --format '{{.ID}} {{.Image}} {{.Names}}'| grep -E "$REGEX" | while read -r i; do
-  CONTAINER="$(echo "$i"|cut -d' ' -f1)"
-  if [ -n "$CONTAINER" ] && [ "${FORCE:-0}" -eq 1 ]; then
-    printf "* Removing existing container '%s'... " "$i"
-    docker rm -f -v "$CONTAINER" 2>/dev/null
-  else
-    echo "WARNING: container '$i' already exists, to remove it: 'FORCE=1 ./docker-run.sh'"
-  fi
-done
+
+if [ "${WEBUI:-0}" -eq 1 ]; then
+  REGEX="$REGEX_WEBUI"
+fi
+
+if [ "${GLFTPD:-0}" -eq 1 ] && [ "${WEBUI:-0}" -eq 1 ]; then
+  REGEX="(${REGEX_GLFTPD}|${REGEX_WEBUI}|${DOCKER_REGISTRY}/docker-glftpd.*)"
+fi
+
+if [ -n "$REGEX" ]; then
+  docker ps -a --format '{{.ID}} {{.Image}} {{.Names}}'| grep -E "$REGEX" | while read -r i; do
+    CONTAINER="$(echo "$i"|cut -d' ' -f1)"
+    if [ -n "$CONTAINER" ] && [ "${FORCE:-0}" -eq 1 ]; then
+      printf "* Removing existing container '%s'... " "$i"
+      docker rm -f -v "$CONTAINER" 2>/dev/null
+    else
+      echo "WARNING: container '$i' already exists, to remove it: 'FORCE=1 ./docker-run.sh'"
+    fi
+  done
+fi
 
 # run docker with glftpd image and GLFTPD_ARGS
 
 # shellcheck disable=SC2086
 if [ "${GLFTPD:-1}" -eq 1 ]; then
+  LOCAL_GLFTPD_IMAGE=$(
+    docker image ls --format='{{.Repository}}' --filter reference="$DOCKER_IMAGE_GLFTPD"
+  )
+  LOCAL_FULL_GLFTPD_IMAGE=$(
+    docker image ls --format='{{.Repository}}' --filter reference="${DOCKER_IMAGE_GLFTPD/%:latest/:full}"
+  )
+  if [ -n "$LOCAL_GLFTPD_IMAGE" ] && [ "${USE_FULL:-0}" -eq 0 ]; then
+    echo "* Found local image 'docker-glftpd'"
+  elif [ -n "$LOCAL_FULL_GLFTPD_IMAGE" ]; then
+    DOCKER_IMAGE_GLFTPD="${DOCKER_IMAGE_GLFTPD/%:latest/:full}"
+    echo "* Using full docker image ${LOCAL_FULL_GLFTPD_IMAGE:-""})"
+  else
+    # check if we already have 'full' tagged image, then keep using it
+    REGISTRY_FULL_GLFTPD_IMAGE=$(
+      docker image ls --format='{{.Repository}}' --filter reference="${DOCKER_REGISTRY}/${DOCKER_IMAGE_GLFTPD/%:latest/:full}"
+    )
+    if [ -n "$REGISTRY_FULL_GLFTPD_IMAGE" ] || [ "${USE_FULL:-0}" -eq 1 ]; then
+      DOCKER_IMAGE_GLFTPD="${DOCKER_IMAGE_GLFTPD/%:latest/:full}"
+    fi
+    echo "* Pulling image '${DOCKER_IMAGE_GLFTPD}' from registry '$DOCKER_REGISTRY'"
+    DOCKER_IMAGE_GLFTPD="${DOCKER_REGISTRY}/$DOCKER_IMAGE_GLFTPD"
+    docker pull "$DOCKER_IMAGE_GLFTPD"
+  fi
   if [ -n "$DOCKER_IMAGE_GLFTPD" ]; then
     printf "* Docker run '%s'... " "$DOCKER_IMAGE_GLFTPD"
     docker run \
       $GLFTPD_ARGS \
-      --detach \
       --name glftpd \
       --hostname glftpd \
       --publish "${IP_ADDR}:${GLFTPD_PORT:-1337}:1337" \
+      --network "${NETWORK:-bridge}" \
       --workdir /glftpd \
       $DOCKER_IMAGE_GLFTPD
     echo "* For logs run 'docker logs glftpd'"
   fi
 fi
 
-# run optional web interface
+# run web interface image with WEBUI_ARGS
 
 if [ "${WEBUI:-1}" -eq 1 ]; then
   LOCAL_IMAGE_WEBUI=$(
@@ -225,17 +267,12 @@ if [ "${WEBUI:-1}" -eq 1 ]; then
   fi
   # shellcheck disable=SC2086
   if [ -n "$DOCKER_IMAGE_WEBUI" ]; then
-    if [ "${RM:-1}" -eq 1 ]; then
-      WEBUI_ARGS+=" --rm  "
-    fi
     printf "* Docker run '%s'... " "$DOCKER_IMAGE_WEBUI"
-    #--detach \
-      
     docker run \
       $WEBUI_ARGS \
       --hostname glftpd-web \
       --name glftpd-web \
-      --publish "${IP_ADDR:-127.0.0.1}:4444:443" \
+      --network "${NETWORK:-bridge}" \
       $DOCKER_IMAGE_WEBUI
   fi
   echo "* For logs run 'docker logs glftpd-web'"
